requirements.txt
fastapi==0.111.0
jinja2==3.1.2
request==2.31.0